requirements.txt
streamlit
google-generativeai
PyPDF2
python-dotenv
streamlit-chat
pandas
